requirements.txt
discord
googletrans==3.1.0a0
PyDictionary
pyinstaller
python-dateutil
